requirements.txt
# Core Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-RESTful>=0.3.10
Werkzeug>=3.0.0

# MCP Server
mcp>=1.0.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.8.6
httpx>=0.25.0

# Data Processing
pydantic>=2.4.0
pydantic-settings>=2.0.0

# File Operations
python-magic>=0.4.27
Pillow>=10.4.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Logging
structlog>=23.0.0
colorlog>=6.7.0
python-json-logger>=2.0.0

# Validation
jsonschema>=4.19.0

# Async Utilities
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development Tools
black>=23.9.0
flake8>=6.1.0

# Security
cryptography>=41.0.0

# WebSocket Support
websockets>=11.0.0

# JSON Processing
orjson>=3.9.0

# Date/Time
python-dateutil>=2.8.0
pytz>=2023.0

# File Upload
python-multipart>=0.0.6

# Caching
redis>=5.0.0

# Background Tasks
celery>=5.3.0

# API Documentation
flasgger>=0.9.7

# Health Checks
healthcheck>=1.3.0 